+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/timeline/item/AbsMessageItem.kt
@@ -162,10 +162,15 @@ abstract class AbsMessageItem<H : AbsMessageItem.Holder> : BaseEventItem<H>() {
         return true
     }
 
-    protected fun renderSendState(root: View, textView: TextView?) {
+    protected open fun renderSendState(root: View, textView: TextView?, failureIndicator: ImageView? = null) {
         root.isClickable = informationData.sendState.isSent()
         val state = if (informationData.hasPendingEdits) SendState.UNSENT else informationData.sendState
         textView?.setTextColor(colorProvider.getMessageTextColor(state))
+        failureIndicator?.isVisible = when (informationData.sendState) {
+            SendState.UNDELIVERED,
+            SendState.FAILED_UNKNOWN_DEVICES -> true
+            else                             -> false
+        }
     }
 
     abstract class Holder(@IdRes stubId: Int) : BaseHolder(stubId) {